+++ src/tests/helpers/quality_of_life.py
@@ -263,6 +263,18 @@ def drain_energy_for_all(game_state: GameState) -> GameState:
     ).build()
 
 
+def recharge_energy_for(game_state: GameState, pid: Pid, char_id: None | int = None) -> GameState:
+    return game_state.factory().f_player(
+        pid,
+        lambda p: p.factory().f_characters(
+            lambda cs: cs.factory().f_character(
+                cs.just_get_active_character_id() if char_id is None else char_id,
+                lambda c: c.factory().energy(c.max_energy).build()
+            ).build()
+        ).build()
+    ).build()
+
+
 def fill_dice_with_omni(game_state: GameState) -> GameState:
     dice = {
         Element.OMNI: BIG_INT,